+++ texar/torch/data/vocabulary_test.py
@@ -1,3 +1,16 @@
+# Copyright 2019 The Texar Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """
 Unit tests for vocabulary related operations.
 """
@@ -40,8 +53,6 @@ class VocabularyTest(unittest.TestCase):
             set(vocab.token_to_id_map_py.keys()),
             set(['word', '词'] + vocab.special_tokens))
 
-        # import pdb
-        # pdb.set_trace()
         # Tests UNK token
         unk_token_id = vocab.map_tokens_to_ids_py(['new'])
         unk_token_text = vocab.map_ids_to_tokens_py(unk_token_id)